Addressed PR feedback
* Removed convert from time
