+++ composer/core/time.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import re
 import textwrap
 import warnings
-from typing import TYPE_CHECKING, Optional, Union
+from typing import TYPE_CHECKING, Union
 
 from composer.core.serializable import Serializable
 from composer.utils.string_enum import StringEnum
@@ -300,386 +300,6 @@ class Time:
             value = int(value)
         return Time(value, unit)
 
-    def convert(
-        self,
-        unit: Union[TimeUnit, str],
-        *,
-        batch_size: Optional[int] = None,
-        drop_last: Optional[bool] = None,
-        dataset_num_samples: Optional[int] = None,
-        dataset_num_tokens: Optional[int] = None,
-        max_training_duration: Optional[Union[str, Time]] = None,
-    ) -> Time:
-        r"""Convert a :class:`Time` instance into the specified ``unit``.
-
-        Parameter ``unit`` is always required. The following table lists the additional required parameters
-        to perform the conversion:
-
-        +-----------------------------------------------------+-------------------------------------------------+-----------------------------------------------------------+-----------------------------------------------------------+-----------------------------------+-----------------------------+
-        | From Unit |:arrow_down:| \\ To Unit |:arrow_right:| | :attr:`~TimeUnit.EPOCH`                         | :attr:`~TimeUnit.BATCH`                                   | :attr:`~TimeUnit.SAMPLE`                                  | :attr:`~TimeUnit.TOKEN`           | :attr:`~TimeUnit.DURATION`  |
-        +-----------------------------------------------------+-------------------------------------------------+-----------------------------------------------------------+-----------------------------------------------------------+-----------------------------------+-----------------------------+
-        | :attr:`~TimeUnit.EPOCH`                             | No required parameters.                         | - ``batch_size``                                          | - ``drop_last``                                           | - ``dataset_num_tokens``          | - ``max_training_duration`` |
-        |                                                     |                                                 | - ``drop_last``                                           | - ``dataset_num_samples``                                 | - ``drop_last`` must be ``False`` |                             |
-        |                                                     |                                                 | - ``dataset_num_samples``                                 | - ``batch_size`` (if ``drop_last`` is ``True``)           |                                   |                             |
-        +-----------------------------------------------------+-------------------------------------------------+-----------------------------------------------------------+-----------------------------------------------------------+-----------------------------------+-----------------------------+
-        | :attr:`~TimeUnit.BATCH`                             | - ``batch_size``                                | No required parameters.                                   | - ``batch_size``                                          | Unsupported conversion.           | - ``max_training_duration`` |
-        |                                                     | - ``drop_last``                                 |                                                           | - ``drop_last``                                           |                                   |                             |
-        |                                                     | - ``dataset_num_samples``                       |                                                           | - ``dataset_num_samples`` (if ``drop_last`` is ``False``) |                                   |                             |
-        +-----------------------------------------------------+-------------------------------------------------+-----------------------------------------------------------+-----------------------------------------------------------+-----------------------------------+-----------------------------+
-        | :attr:`~TimeUnit.SAMPLE`                            | - ``drop_last``                                 | - ``batch_size``                                          | No required parameters.                                   | Unsupported conversion.           | - ``max_training_duration`` |
-        |                                                     | - ``dataset_num_samples``                       | - ``drop_last``                                           |                                                           |                                   |                             |
-        |                                                     | - ``batch_size`` (if ``drop_last`` is ``True``) | - ``dataset_num_samples`` (if ``drop_last`` is ``False``) |                                                           |                                   |                             |
-        +-----------------------------------------------------+-------------------------------------------------+-----------------------------------------------------------+-----------------------------------------------------------+-----------------------------------+-----------------------------+
-        | :attr:`~TimeUnit.TOKEN`                             | - ``dataset_num_tokens``                        | Unsupported conversion.                                   | Unsupported conversion.                                   | No required parameters.           | - ``max_training_duration`` |
-        |                                                     | - ``drop_last`` must be ``False``               |                                                           |                                                           |                                   |                             |
-        +-----------------------------------------------------+-------------------------------------------------+-----------------------------------------------------------+-----------------------------------------------------------+-----------------------------------+-----------------------------+
-        | :attr:`~TimeUnit.DURATION`                          | - ``max_training_duration``                     | - ``max_training_duration``                               | - ``max_training_duration``                               | - ``max_training_duration``       | No required parameters.     |
-        +-----------------------------------------------------+-------------------------------------------------+-----------------------------------------------------------+-----------------------------------------------------------+-----------------------------------+-----------------------------+
-
-        For example:
-
-        >>> Time(2, "ep").convert(TimeUnit.BATCH, dataset_num_samples=100, batch_size=50, drop_last=True) == Time("4ba")
-
-        Args:
-            unit (Union[TimeUnit, str]): The desired unit to convert the time instance into.
-            batch_size (int, optional): The optimization batch size.
-            drop_last (bool, optional): Whether the dataloader is dropping last (incomplete) batches.
-            dataset_num_samples (int, optional): The number of samples in the dataset.
-            dataset_num_tokens (int, optional): The number of tokens in the dataset. Required only if
-                converting to or from :attr:`TimeUnit.TOKEN`.
-            max_training_duration (str or Time, optional): The total training duration. Required only
-                if converting to or from :attr:`TimeUnit.DURATION`.
-        
-        Raises:
-            ValueError: If it is not possible to perform the conversion. 
-
-        Returns:
-            Time: The time, in the specified ``unit``.
-        """
-        unit = TimeUnit(unit)
-
-        if unit == self.unit:
-            # No conversion required
-            return Time(self.value, self.unit)
-
-        if unit == TimeUnit.DURATION or self.unit == TimeUnit.DURATION:
-            # if the desired unit is duration, then the logic is the same regardless of the from unit
-            if max_training_duration is None:
-                raise ValueError("max_training_duration is required to convert to or from DURATION")
-            if unit == TimeUnit.DURATION:
-                return self.convert_to_duration(max_training_duration=max_training_duration)
-            else:
-                return self.convert_from_duration(max_training_duration=max_training_duration)
-
-        if self.unit == TimeUnit.EPOCH:
-            if unit == TimeUnit.BATCH:
-                if batch_size is None:
-                    raise ValueError("batch_size is required to convert from EPOCH to BATCH")
-                if drop_last is None:
-                    raise ValueError("drop_last is required to convert from EPOCH to BATCH")
-                if dataset_num_samples is None:
-                    raise ValueError("dataset_num_samples is required to convert from EPOCH to BATCH")
-                return self.convert_epoch_to_batch(batch_size=batch_size,
-                                                   drop_last=drop_last,
-                                                   dataset_num_samples=dataset_num_samples)
-            if unit == TimeUnit.SAMPLE:
-                if dataset_num_samples is None:
-                    raise ValueError("dataset_num_samples is required to convert from EPOCH to SAMPLE")
-                if drop_last is None:
-                    raise ValueError("drop_last is required to convert from EPOCH to SAMPLE")
-                return self.convert_epoch_to_sample(drop_last=drop_last,
-                                                    dataset_num_samples=dataset_num_samples,
-                                                    batch_size=batch_size)
-            if unit == TimeUnit.TOKEN:
-                if dataset_num_tokens is None:
-                    raise ValueError("dataset_num_tokens is required to convert from EPOCH to TOKEN")
-                return self.convert_epoch_to_token(dataset_num_tokens=dataset_num_tokens)
-        if self.unit == TimeUnit.BATCH:
-            if unit == TimeUnit.EPOCH:
-                if batch_size is None:
-                    raise ValueError("batch_size is required to convert from EPOCH to BATCH")
-                if drop_last is None:
-                    raise ValueError("drop_last is required to convert from EPOCH to BATCH")
-                if dataset_num_samples is None:
-                    raise ValueError("dataset_num_samples is required to convert from EPOCH to BATCH")
-                return self.convert_batch_to_epoch(batch_size=batch_size,
-                                                   drop_last=drop_last,
-                                                   dataset_num_samples=dataset_num_samples)
-            if unit == TimeUnit.SAMPLE:
-                if batch_size is None:
-                    raise ValueError("batch_size is required to convert from BATCH to SAMPLE")
-                if drop_last is None:
-                    raise ValueError("drop_last is required to convert from BATCH to SAMPLE")
-                return self.convert_batch_to_sample(batch_size=batch_size,
-                                                    drop_last=drop_last,
-                                                    dataset_num_samples=dataset_num_samples)
-        if self.unit == TimeUnit.SAMPLE:
-            if unit == TimeUnit.EPOCH:
-                if drop_last is None:
-                    raise ValueError("drop_last is required to convert from SAMPLE to EPOCH")
-                if dataset_num_samples is None:
-                    raise ValueError("dataset_num_samples is required to convert from SAMPLE to SAMPLE")
-                return self.convert_sample_to_epoch(drop_last=drop_last,
-                                                    dataset_num_samples=dataset_num_samples,
-                                                    batch_size=batch_size)
-            if unit == TimeUnit.BATCH:
-                if drop_last is None:
-                    raise ValueError("drop_last is required to convert from BATCH to SAMPLE")
-                if batch_size is None:
-                    raise ValueError("batch_size is required to convert from BATCH to SAMPLE")
-                return self.convert_sample_to_batch(batch_size=batch_size,
-                                                    drop_last=drop_last,
-                                                    dataset_num_samples=dataset_num_samples)
-        if self.unit == TimeUnit.TOKEN:
-            if unit == TimeUnit.EPOCH:
-                if dataset_num_tokens is None:
-                    raise ValueError("dataset_num_tokens is required to convert from TOKEN to EPOCH")
-                return self.convert_token_to_epoch(dataset_num_tokens=dataset_num_tokens)
-
-        raise ValueError(f"Unable to convert from {self.unit} to {unit}")
-
-    def convert_epoch_to_batch(self, *, batch_size: int, drop_last: bool, dataset_num_samples: int) -> Time:
-        """Convert ``self`` into :attr:`TimeUnit.BATCH`. Requires that ``self.unit == TimeUnit.EPOCH``.
-
-        Args:
-            batch_size (int): The optimization batch size.
-            drop_last (bool): Whether the dataloader is dropping last (incomplete) batches.
-            dataset_num_samples (int): The number of samples in the dataset.
-
-        Raises:
-            RuntimeError: Raised if ``self.unit != TimeUnit.EPOCH``
-
-        Returns:
-            Time: The time, in :attr:`TimeUnit.BATCH`.
-        """
-        if self.unit != TimeUnit.EPOCH:
-            raise RuntimeError(f"Time {self} units are not epochs.")
-        if drop_last:
-            num_batches_per_epoch = dataset_num_samples // batch_size
-        else:
-            num_batches_per_epoch = (dataset_num_samples - 1) // batch_size + 1
-        return Time(self.value * num_batches_per_epoch, TimeUnit.BATCH)
-
-    def convert_epoch_to_sample(self,
-                                *,
-                                drop_last: bool,
-                                dataset_num_samples: int,
-                                batch_size: Optional[int] = None) -> Time:
-        """Convert ``self`` into :attr:`TimeUnit.SAMPLE`. Requires that ``self.unit == TimeUnit.EPOCH``.
-
-        Args:
-            drop_last (bool): Whether the dataloader is dropping last (incomplete) batches.
-            dataset_num_samples (int): The number of samples in the dataset.
-            batch_size (int, optional): The optimization batch size. Required if ``drop_last`` is ``True``.
-
-        Raises:
-            RuntimeError: Raised if ``self.unit != TimeUnit.EPOCH``
-
-        Returns:
-            Time: The time, in :attr:`TimeUnit.SAMPLE`.
-        """
-        if self.unit != TimeUnit.EPOCH:
-            raise RuntimeError(f"Time {self} units are not epochs.")
-        if drop_last:
-            if batch_size is None:
-                raise ValueError("batch_size is required to convert from EPOCH to SAMPLE when drop_last=True")
-            num_batches_per_epoch = dataset_num_samples // batch_size
-            num_samples_per_epoch = num_batches_per_epoch * batch_size
-            return Time(self.value * num_samples_per_epoch, TimeUnit.SAMPLE)
-        else:
-            return Time(self.value * dataset_num_samples, TimeUnit.SAMPLE)
-
-    def convert_epoch_to_token(self, *, dataset_num_tokens: int) -> Time:
-        """Convert ``self`` into :attr:`TimeUnit.TOKEN`. Requires that ``self.unit == TimeUnit.EPOCH``.
-
-        .. note::
-
-            The conversion is valid only if the dataloader yields all batches (i.e. ``drop_last`` is ``False``).
-
-        Args:
-            dataset_num_tokens (int): The number of tokens in the dataset.
-
-        Raises:
-            RuntimeError: Raised if ``self.unit != TimeUnit.EPOCH``
-
-        Returns:
-            Time: The time, in :attr:`TimeUnit.TOKEN`.
-        """
-        if self.unit != TimeUnit.EPOCH:
-            raise RuntimeError(f"Time {self} units are not epochs.")
-        return Time(self.value * dataset_num_tokens, TimeUnit.TOKEN)
-
-    def convert_batch_to_epoch(self, *, batch_size: int, drop_last: bool, dataset_num_samples: int):
-        """Convert ``self`` into :attr:`TimeUnit.EPOCH`. Requires that ``self.unit == TimeUnit.BATCH``.
-
-        Args:
-            batch_size (int): The optimization batch size.
-            drop_last (bool): Whether the dataloader is dropping last (incomplete) batches.
-            dataset_num_samples (int): The number of samples in the dataset.
-
-        Raises:
-            RuntimeError: Raised if ``self.unit != TimeUnit.BATCH``
-
-        Returns:
-            Time: The time, in :attr:`TimeUnit.EPOCH`.
-        """
-        if self.unit != TimeUnit.BATCH:
-            raise RuntimeError(f"Time {self} units are not batches.")
-        if drop_last:
-            num_batches_per_epoch = dataset_num_samples // batch_size
-        else:
-            num_batches_per_epoch = (dataset_num_samples - 1) // batch_size + 1
-        return Time(self.value // num_batches_per_epoch, TimeUnit.EPOCH)
-
-    def convert_batch_to_sample(
-        self,
-        *,
-        batch_size: int,
-        drop_last: bool,
-        dataset_num_samples: Optional[int] = None,
-    ) -> Time:
-        """Convert ``self`` into :attr:`TimeUnit.SAMPLE`. Requires that ``self.unit == TimeUnit.BATCH``.
-
-        Args:
-            batch_size (int): The optimization batch size.
-            drop_last (bool): Whether the dataloader is dropping last (incomplete) batches.
-            dataset_num_samples (int, optional): The number of samples in the dataset. Required if ``drop_last`` is ``False``.
-
-        Raises:
-            RuntimeError: Raised if ``self.unit != TimeUnit.BATCH``
-
-        Returns:
-            Time: The time, in :attr:`TimeUnit.SAMPLE`.
-        """
-        if self.unit != TimeUnit.BATCH:
-            raise RuntimeError(f"Time {self} units are not batches.")
-        if drop_last:
-            return Time(self.value * batch_size, TimeUnit.SAMPLE)
-        else:
-            if dataset_num_samples is None:
-                raise ValueError("dataset_num_samples is required to convert from EPOCH to BATCH when drop_last=False")
-            num_batches_per_epoch = (dataset_num_samples - 1) // batch_size + 1
-            samples_from_complete_epochs = (self.value // num_batches_per_epoch) * dataset_num_samples
-            remaining_batches = self.value % num_batches_per_epoch
-            samples_from_incomplete_epochs = remaining_batches * batch_size
-            return Time(samples_from_complete_epochs + samples_from_incomplete_epochs, TimeUnit.SAMPLE)
-
-    def convert_sample_to_epoch(self,
-                                *,
-                                drop_last: bool,
-                                dataset_num_samples: int,
-                                batch_size: Optional[int] = None) -> Time:
-        """Convert ``self`` into :attr:`TimeUnit.EPOCH`. Requires that ``self.unit == TimeUnit.SAMPLE``.
-
-        Args:
-            drop_last (bool): Whether the dataloader is dropping last (incomplete) batches.
-            dataset_num_samples (int): The number of samples in the dataset.
-            batch_size (int, optional): The optimization batch size. Required if ``drop_last`` is ``True``.
-
-        Raises:
-            RuntimeError: Raised if ``self.unit != TimeUnit.SAMPLE``
-
-        Returns:
-            Time: The time, in :attr:`TimeUnit.EPOCH`.
-        """
-        if self.unit != TimeUnit.SAMPLE:
-            raise RuntimeError(f"Time {self} units are not samples.")
-        if drop_last:
-            if batch_size is None:
-                raise ValueError("batch_size is required to convert from SAMPLE to SAMPLE when drop_last=True")
-            num_batches_per_epoch = dataset_num_samples // batch_size
-            num_samples_per_epoch = num_batches_per_epoch * batch_size
-            return Time(self.value // num_samples_per_epoch, TimeUnit.EPOCH)
-        else:
-            return Time(self.value // dataset_num_samples, TimeUnit.EPOCH)
-
-    def convert_sample_to_batch(
-        self,
-        *,
-        batch_size: int,
-        drop_last: bool,
-        dataset_num_samples: Optional[int] = None,
-    ) -> Time:
-        """Convert ``self`` into :attr:`TimeUnit.BATCH`. Requires that ``self.unit == TimeUnit.SAMPLE``.
-
-        Args:
-            batch_size (int): The optimization batch size.
-            drop_last (bool): Whether the dataloader is dropping last (incomplete) batches.
-            dataset_num_samples (int, optional): The number of samples in the dataset. Required if ``drop_last`` is ``False``.
-
-        Raises:
-            RuntimeError: Raised if ``self.unit != TimeUnit.SAMPLE``
-
-        Returns:
-            Time: The time, in :attr:`TimeUnit.BATCH`.
-        """
-        if self.unit != TimeUnit.SAMPLE:
-            raise RuntimeError(f"Time {self} units are not samples.")
-        if drop_last:
-            return Time(self.value // batch_size, TimeUnit.BATCH)
-        else:
-            if dataset_num_samples is None:
-                raise ValueError("dataset_num_samples is required to convert from SAMPLE to BATCH when drop_last=False")
-            num_batches_per_epoch = (dataset_num_samples - 1) // batch_size + 1
-            batches_from_complete_epochs = (self.value // dataset_num_samples) * num_batches_per_epoch
-            batches_from_incomplete_epoch = (self.value % dataset_num_samples) // batch_size
-            return Time(batches_from_complete_epochs + batches_from_incomplete_epoch, TimeUnit.BATCH)
-
-    def convert_token_to_epoch(self, *, dataset_num_tokens: int) -> Time:
-        """Convert ``self`` into :attr:`TimeUnit.EPOCH`. Requires that ``self.unit == TimeUnit.TOKEN``.
-
-        .. note::
-
-            The conversion is valid only if the dataloader yields all batches (i.e. ``drop_last`` == ``False``).
-
-        Args:
-            dataset_num_tokens (int): The number of tokens in the dataset.
-
-        Raises:
-            RuntimeError: Raised if ``self.unit != TimeUnit.TOKEN``
-
-        Returns:
-            Time: The time, in :attr:`TimeUnit.EPOCH`.
-        """
-        if self.unit != TimeUnit.TOKEN:
-            raise RuntimeError(f"Time {self} units are not tokens.")
-        return Time(self.value // dataset_num_tokens, TimeUnit.EPOCH)
-
-    def convert_to_duration(self, *, max_training_duration: Union[str, Time]) -> Time:
-        """Convert ``self`` into :attr:`TimeUnit.DURATION`.
-
-        Args:
-            max_training_duration (str or Time): The total training duration.
-
-        Returns:
-            Time: The time, in :attr:`TimeUnit.DURATION`.
-        """
-        if self.unit == TimeUnit.DURATION:
-            return Time(self.value, self.unit)
-        if isinstance(max_training_duration, str):
-            max_training_duration = Time.from_timestring(max_training_duration)
-        return self / max_training_duration
-
-    def convert_from_duration(self, *, max_training_duration: Union[str, Time]) -> Time:
-        """Convert ``self`` into the units of ``max_training_duration``.
-
-        Args:
-            max_training_duration (str or Time): The total training duration.
-
-        Raises:
-            RuntimeError: Raised if ``self.unit != TimeUnit.DURATION``
-
-        Returns:
-            Time: The time, in the units of ``max_training_duration``.
-        """
-        if self.unit != TimeUnit.DURATION:
-            raise RuntimeError(f"Time {self} units is not duration.")
-        if isinstance(max_training_duration, str):
-            max_training_duration = Time.from_timestring(max_training_duration)
-        if max_training_duration.unit == TimeUnit.DURATION:
-            return Time(self.value, self.unit)
-        return self * max_training_duration
-
 
 class Timer(Serializable):
     """Timer tracks the current training progress, in terms of epochs, batches, samples, and tokens."""
@@ -723,6 +343,26 @@ class Timer(Serializable):
     def token(self) -> Time:
         """The current token."""
         return self._token
+
+    def get(self, unit: Union[str, TimeUnit]) -> Time:
+        """Returns the current time in the specified unit.
+
+        Args:
+            unit (str or TimeUnit): The desired unit.
+
+        Returns:
+            Time: The current time, in the specified unit.
+        """
+        unit = TimeUnit(unit)
+        if unit == TimeUnit.EPOCH:
+            return self.epoch
+        if unit == TimeUnit.BATCH:
+            return self.batch
+        if unit == TimeUnit.SAMPLE:
+            return self.sample
+        if unit == TimeUnit.TOKEN:
+            return self.token
+        raise ValueError(f"Invalid unit: {unit}")
 
     def on_batch_complete(self, samples: Union[int, Time] = 0, tokens: Union[int, Time] = 0):
         """Called by the trainer at the end of every optimization batch.
